Re-add service-env identity storaage with build tag
