+++ src/internal/serviceenv/service_env_cgo.go
@@ -1,0 +1,5 @@
+// +build cgo
+
+package serviceenv
+
+func (env *NonblockingServiceEnv) InitDexDB() {}
